+++ src/main/java/com/github/kd_gaming1/packcore/config/PackCoreConfig.java
@@ -70,6 +70,10 @@ public class PackCoreConfig extends MidnightConfig {
 
     @Hidden
     @Entry(category = META)
+    public static String lastSeenModpackVersion = "";
+
+    @Hidden
+    @Entry(category = META)
     public static long lastBackupEpochMs = 0L;
 
     @Hidden